Initial implementation for cross-platform server
