+++ src/ImoutoDesktop.Server/Program.cs
@@ -15,9 +15,20 @@ namespace ImoutoDesktop.Server
             var port = 1024;
             var ipAddress = (await Dns.GetHostAddressesAsync(Dns.GetHostName())).First(x => x.AddressFamily == AddressFamily.InterNetwork);
 
+            var serviceImpl = default(RemoteServiceImpl);
+
+            if (Environment.OSVersion.Platform == PlatformID.Win32NT)
+            {
+                serviceImpl = new WindowsRemoteServiceImpl();
+            }
+            else
+            {
+                serviceImpl = new UnixRemoteServiceImpl();
+            }
+
             var server = new Grpc.Core.Server
             {
-                Services = { Remoting.RemoteService.BindService(new RemoteServiceImpl()) },
+                Services = { Remoting.RemoteService.BindService(serviceImpl) },
                 Ports = { new ServerPort("0.0.0.0", port, ServerCredentials.Insecure) }
             };
